+++ CHANGES.md
@@ -5,7 +5,7 @@ Release Notes.
 Apollo 3.0.0
 
 ------------------
-* 
+* [Fix: include super admin in hasAnyPermission semantics](https://github.com/apolloconfig/apollo/pull/5568)
 
 ------------------
 All issues and pull requests are [here](https://github.com/apolloconfig/apollo/milestone/18?closed=1)